phase5_llm/requirements.txt
# Phase 5: LLM (GROQ)
groq>=0.4.0
python-dotenv>=1.0.0
